\name{sample_regular}

\alias{sample_regular}

\title{
Regular sample of a vector
}

\description{
Take a regular sample of a vector
}


\usage{
sample_regular(x, n, sorted=FALSE)
}
 
\arguments{
  \item{x}{vector}
  \item{n}{postive integer. The number of samples}
  \item{sorted}{logical. Should \code{x} be sorted before sampling?}
}

\value{
vector
}

